+++ creational/chain-of-responsibility/src/Descontos/DescontoMaisDe500Reais.php
@@ -6,7 +6,7 @@ use Alura\DesignPattern\Orcamento;
 
 class DescontoMaisDe500Reais extends Desconto
 {
-    public function calculaDesconto(Orcamento $orcamento): float
+    public function calculaDesconto(Orcamento $orcamento): int
     {
         if ($orcamento->valor <= 500) {
             return $orcamento->valor * 0.05;
